Fix FullCalendar column sizing when view is embedded in node body
The main issue was that `max-width: none;` was being overridden by an
overly general, high priority selector.  Temporarily adding the !important
flag until we get a chance to do a proper fix and remove the smelly code
that was interfering with this.  Moved customizations for the view's page
display to its own block.